Move OAuthBase to OAuth namespace
This completes the refactoring.
